some renames
cleaning up and such...
