+++ package.json
@@ -14,8 +14,8 @@
     "dev": "echo 'removing previous /dist folder and some generated public files' && rm -rf ./dist && npm run clean && node --no-experimental-require-module --experimental-print-required-tla ./node_modules/.bin/webpack serve --hot --port=3000"
   },
   "dependencies": {
-    "@babel/core": "^7.26.10",
-    "@babel/preset-env": "^7.26.9",
+    "@babel/core": "^7.27.4",
+    "@babel/preset-env": "^7.27.2",
     "@mdit/plugin-abbr": "^0.16.0",
     "@mdit/plugin-container": "^0.16.0",
     "@mdit/plugin-footnote": "^0.16.8",
@@ -25,13 +25,13 @@
     "@mdit/plugin-sub": "^0.16.0",
     "@mdit/plugin-sup": "^0.16.0",
     "@mdit/plugin-tasklist": "^0.16.0",
-    "autoprefixer": "^10.4.20",
+    "autoprefixer": "^10.4.21",
     "babel-loader": "^9.2.1",
     "bulma": "^1.0.4",
     "copy-webpack-plugin": "^12.0.2",
     "core-js": "3.39.0",
     "css-loader": "^7.1.2",
-    "dotenv": "^16.4.7",
+    "dotenv": "^16.5.0",
     "feed": "^4.2.2",
     "fontawesome-free": "^1.0.4",
     "highlight.js": "^11.11.1",
@@ -47,12 +47,12 @@
     "pug": "^3.0.3",
     "raw-loader": "^4.0.2",
     "regenerator-runtime": "^0.14.1",
-    "sass": "^1.86.3",
+    "sass": "^1.89.2",
     "sass-loader": "^16.0.5",
     "style-loader": "^4.0.0",
-    "tocbot": "^4.36.3",
-    "webpack": "^5.99.5",
+    "tocbot": "^4.36.4",
+    "webpack": "^5.99.9",
     "webpack-cli": "^5.1.4",
-    "webpack-dev-server": "^5.2.1"
+    "webpack-dev-server": "^5.2.2"
   }
 }
