--- support/manifest.js
@@ -1,20 +1,0 @@
-export default {
-  short_name: 'castle',
-  name: 'castle ',
-  icons: [
-    {
-      src: '/favicon-192x192.png',
-      sizes: '192x192',
-      type: 'image/png',
-    },
-    {
-      src: '/favicon-512x512.png',
-      sizes: '512x512',
-      type: 'image/png',
-    },
-  ],
-  start_url: '.',
-  display: 'standalone',
-  theme_color: '#000000',
-  background_color: '#ffffff',
-}
